spin_doctor
Trait Diagnosis

pub trait Diagnosis:
    Debug
    + Send
    + Sync
    + 'static {
    // Required method
    fn description(&self) -> String;

    // Provided methods
    fn is_critical(&self) -> bool { ... }
    fn treatment(&self) -> Option<&dyn Treatment> { ... }
}

The Diagnosis trait represents a detected problem with a Spin app.

Required Methods

fn description(&self) -> String

Return a human-friendly description of this problem.

Provided Methods

fn is_critical(&self) -> bool

Return true if this problem is “critical”, i.e. if the app’s configuration or environment is invalid. Return false for “non-critical” problems like deprecations.

fn treatment(&self) -> Option<&dyn Treatment>

Return a Treatment that can (potentially) fix this problem, or None if there is no automatic fix.