Crate spin_doctor

Spin doctor: check and automatically fix problems with Spin apps.

Modules

manifest

Diagnoses for app manifest format problems.

rustlang

Diagnose for Rust-specific problems.

wasm

Diagnoses for Wasm source problems.

Structs

Checkup

Configuration for an app to be checked for problems.

DryRunNotSupported

Error returned by Treatment::dry_run if dry run isn’t supported.

PatientApp

An app “patient” to be checked for problems.

PatientDiagnosis

A PatientDiagnosis bundles a Diagnosis with its (borrowed) PatientApp.

StopDiagnosing

Return this as an error from a treatment to stop further diagnoses when the user needs to intervene before the doctor can proceed.

Traits

Diagnosis

The Diagnosis trait represents a detected problem with a Spin app.

Diagnostic

The Diagnose trait implements the detection of a particular Spin app problem.

Treatment

The Treatment trait represents a (potential) fix for a detected problem.